move folder chart to inside folder graphic chart, change little style and position some components
